fix structures of map components
